+++ CHANGELOG.md
@@ -22,7 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - adapters: add an input connector for Amazon S3 (#1485)
 - SQL: support for ARRAY_REPEAT function
   ([#1497](https://github.com/feldera/feldera/pull/1497))
-
+- SQL: support for ROUND(DOUBLE, digit), TRUNCATE(DOUBLE, digit)
+  ([#1512](https://github.com/feldera/feldera/pull/1512))
+  
 ## [0.11.0] - 2024-03-05
 
 ### Added